Address review: key hook dedup per concrete class; extract shared hook-manager reset trait
A subclass keeping a core gateway id must not have its callbacks suppressed
by the core instance registering first (or vice versa), so the once-only key
now includes the concrete class. The hook-manager singleton reset duplicated
across three test classes moves to a shared trait.
