+++ includes/payment-methods/class-wc-stripe-upe-payment-method.php
@@ -201,7 +201,10 @@ abstract class WC_Stripe_UPE_Payment_Method extends WC_Payment_Gateway {
 			return;
 		}
 
-		if ( $hook_manager->register_payment_method_hooks( $this->id, WC_Stripe_Hook_Categories::GENERAL ) ) {
+		// Key on the concrete class too: a third-party subclass that keeps a
+		// core id must still register its (possibly overridden) callbacks, and
+		// must not suppress the core instance's — whichever constructs first.
+		if ( $hook_manager->register_payment_method_hooks( $this->id . ':' . get_class( $this ), WC_Stripe_Hook_Categories::GENERAL ) ) {
 			$register();
 		}
 	}
